+++ Sources/Endpoints/Comments.swift
@@ -17,9 +17,8 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect retrieval.
     ///
     /// - important: It requires *public_content* scope for media that does not belong to your own user.
-
     public func comments(fromMedia mediaId: String, success: SuccessHandler<[InstagramComment]>?, failure: FailureHandler?) {
-        request("/media/\(mediaId)/comments", success: success, failure: failure)
+        request("/media/\(mediaId)/comments", success: { data in success?(data!) }, failure: failure)
     }
 
     /// Create a comment on a media object.
@@ -28,21 +27,15 @@ extension Instagram {
     /// - parameter text: Text to post as a comment on the media object as specified in `mediaId`.
     /// - parameter failure: The callback called after an incorrect creation.
     ///
-    /// - important: It requires *comments* scope. Also, *public_content* scope is required for media that does not
-    ///   belong to your own user.
+    /// - important: It requires *comments* scope. Also, *public_content* scope is required for media that does not belong to your own user.
     ///
     /// - note:
     ///     - The total length of the comment cannot exceed 300 characters.
     ///     - The comment cannot contain more than 4 hashtags.
     ///     - The comment cannot contain more than 1 URL.
     ///     - The comment cannot consist of all capital letters.
-
-    public func createComment(onMedia mediaId: String, text: String, failure: FailureHandler?) {
-        var parameters = Parameters()
-
-        parameters["text"] = text
-
-        request("/media/\(mediaId)/comments", method: .post, parameters: parameters, success: { (_: InstagramResponse<Any?>) in return }, failure: failure)
+    public func createComment(onMedia mediaId: String, text: String, success: SuccessHandler<InstagramComment>?, failure: FailureHandler?) {
+        request("/media/\(mediaId)/comments", method: .post, parameters: ["text": text], success: { data in success?(data!) }, failure: failure)
     }
 
     /// Remove a comment either on the authenticated user's media object or authored by the authenticated user.
@@ -51,11 +44,8 @@ extension Instagram {
     /// - parameter mediaId: The ID of the media object to reference.
     /// - parameter failure: The callback called after an incorrect deletion.
     ///
-    /// - important: It requires *comments* scope. Also, *public_content* scope is required for media that does not
-    ///   belong to your own user.
-
-    public func deleteComment(_ commentId: String, onMedia mediaId: String, failure: FailureHandler?) {
-        request("/media/\(mediaId)/comments/\(commentId)", method: .delete, success: { (_: InstagramResponse<Any?>) in return }, failure: failure)
+    /// - important: It requires *comments* scope. Also, *public_content* scope is required for media that does not belong to your own user.
+    public func deleteComment(_ commentId: String, onMedia mediaId: String, success: EmptySuccessHandler?, failure: FailureHandler?) {
+        request("/media/\(mediaId)/comments/\(commentId)", method: .delete, success: { (_: InstagramEmptyResponse!) in success?() }, failure: failure)
     }
-
 }