+++ Sources/Endpoints/Likes.swift
@@ -17,9 +17,8 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect retrieval.
     ///
     /// - important: It requires *public_content* scope for media that does not belong to your own user.
-
     public func likes(inMedia mediaId: String, success: SuccessHandler<[InstagramUser]>?, failure: FailureHandler?) {
-        request("/media/\(mediaId)/likes", success: success, failure: failure)
+        request("/media/\(mediaId)/likes", success: { data in success?(data!) }, failure: failure)
     }
 
     /// Set a like on this media by the currently authenticated user.
@@ -27,11 +26,9 @@ extension Instagram {
     /// - parameter mediaId: The ID of the media object to reference.
     /// - parameter failure: The callback called after an incorrect like.
     ///
-    /// - important: It requires *likes* scope. Also, *public_content* scope is required for media that does not belong
-    ///   to your own user.
-
-    public func like(media mediaId: String, failure: FailureHandler?) {
-        request("/media/\(mediaId)/likes", method: .post, success: { (_: InstagramResponse<Any?>) in return }, failure: failure)
+    /// - important: It requires *likes* scope. Also, *public_content* scope is required for media that does not belong to your own user.
+    public func like(media mediaId: String, success: EmptySuccessHandler?, failure: FailureHandler?) {
+        request("/media/\(mediaId)/likes", method: .post, success: { (_: InstagramEmptyResponse!) in success?() }, failure: failure)
     }
 
     /// Remove a like on this media by the currently authenticated user.
@@ -39,11 +36,8 @@ extension Instagram {
     /// - parameter Parameter mediaId: The ID of the media object to reference.
     /// - parameter failure: The callback called after an incorrect deletion.
     ///
-    /// - important: It requires *likes* scope. Also, *public_content* scope is required for media that does not belong
-    ///   to your own user.
-
-    public func unlike(media mediaId: String, failure: FailureHandler?) {
-        request("/media/\(mediaId)/likes", method: .delete, success: { (_: InstagramResponse<Any?>) in return }, failure: failure)
+    /// - important: It requires *likes* scope. Also, *public_content* scope is required for media that does not belong to your own user.
+    public func unlike(media mediaId: String, success: EmptySuccessHandler?, failure: FailureHandler?) {
+        request("/media/\(mediaId)/likes", method: .delete, success: { (_: InstagramEmptyResponse!) in success?() }, failure: failure)
     }
-
 }