+++ Sources/Endpoints/Relationships.swift
@@ -11,7 +11,6 @@ extension Instagram {
     // MARK: - Relationship Endpoints
 
     /// Relationship actions currently supported by Instagram.
-
     private enum RelationshipAction: String {
         case follow, unfollow, approve, ignore
     }
@@ -22,9 +21,8 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect retrieval.
     ///
     /// - important: It requires *follower_list* scope.
-
     public func userFollows(success: SuccessHandler<[InstagramUser]>?, failure: FailureHandler?) {
-        request("/users/self/follows", success: success, failure: failure)
+        request("/users/self/follows", success: { data in success?(data!) }, failure: failure)
     }
 
     /// Get the list of users this user is followed by.
@@ -33,9 +31,8 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect retrieval.
     ///
     /// - important: It requires *follower_list* scope.
-
     public func userFollowers(success: SuccessHandler<[InstagramUser]>?, failure: FailureHandler?) {
-        request("/users/self/followed-by", success: success, failure: failure)
+        request("/users/self/followed-by", success: { data in success?(data!) }, failure: failure)
     }
 
     /// List the users who have requested this user's permission to follow.
@@ -44,9 +41,8 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect retrieval.
     ///
     /// - important: It requires *follower_list* scope.
-
     public func userRequestedBy(success: SuccessHandler<[InstagramUser]>?, failure: FailureHandler?) {
-        request("/users/self/requested-by", success: success, failure: failure)
+        request("/users/self/requested-by", success: { data in success?(data!) }, failure: failure)
     }
 
     /// Get information about a relationship to another user.
@@ -56,9 +52,8 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect retrieval.
     ///
     /// - important: It requires *follower_list* scope.
-
     public func userRelationship(withUser userId: String, success: SuccessHandler<InstagramRelationship>?, failure: FailureHandler?) {
-        request("/users/\(userId)/relationship", success: success, failure: failure)
+        request("/users/\(userId)/relationship", success: { data in success?(data!) }, failure: failure)
     }
 
     /// Modify the relationship between the current user and the target user.
@@ -69,16 +64,12 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect modification.
     ///
     /// - important: It requires *relationships* scope.
-
     private func modifyUserRelationship(withUser userId: String,
                                         action: RelationshipAction,
                                         success: SuccessHandler<InstagramRelationship>?,
                                         failure: FailureHandler?) {
-        var parameters = Parameters()
 
-        parameters["action"] = action.rawValue
-
-        request("/users/\(userId)/relationship", method: .post, parameters: parameters, success: success, failure: failure)
+        request("/users/\(userId)/relationship", method: .post, parameters: ["action": action.rawValue], success: { data in success?(data!) }, failure: failure)
     }
 
     /// Follows the target user.
@@ -88,7 +79,6 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect follow.
     ///
     /// - important: It requires *relationships* scope.
-
     public func follow(user userId: String, success: SuccessHandler<InstagramRelationship>?, failure: FailureHandler?) {
         modifyUserRelationship(withUser: userId, action: .follow, success: success, failure: failure)
     }
@@ -100,7 +90,6 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect unfollow.
     ///
     /// - important: It requires *relationships* scope.
-
     public func unfollow(user userId: String, success: SuccessHandler<InstagramRelationship>?, failure: FailureHandler?) {
         modifyUserRelationship(withUser: userId, action: .unfollow, success: success, failure: failure)
     }
@@ -112,7 +101,6 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect approve.
     ///
     /// - important: It requires *relationships* scope.
-
     public func approveRequest(fromUser userId: String, success: SuccessHandler<InstagramRelationship>?, failure: FailureHandler?) {
         modifyUserRelationship(withUser: userId, action: .approve, success: success, failure: failure)
     }
@@ -124,9 +112,7 @@ extension Instagram {
     /// - parameter failure: The callback called after an incorrect ignore.
     ///
     /// - important: It requires *relationships* scope.
-
     public func ignoreRequest(fromUser userId: String, success: SuccessHandler<InstagramRelationship>?, failure: FailureHandler?) {
         modifyUserRelationship(withUser: userId, action: .ignore, success: success, failure: failure)
     }
-
 }